+++ src/h/fossil_engine.h
@@ -36,7 +36,6 @@
 #include "shaders.h"
 #include "string.h"
 #include "super_debugger.h"
-#include "text.h"
 #include "time.h"
 #include "ui.h"