+++ tests/composable_ui/src/main.c
@@ -2,9 +2,7 @@
 
 int main(int argc, char **argv)
 {
-    if (
-            fsl_engine_init(argc, argv, NULL, NULL, 1280, 720, NULL,
-                FSL_FLAG_LOAD_DEFAULT_SHADERS) != FSL_ERR_SUCCESS ||
+    if (fsl_engine_init(argc, argv, NULL, NULL, 1280, 720, NULL, 0) != FSL_ERR_SUCCESS ||
             fsl_ui_init(FALSE) != FSL_ERR_SUCCESS)
         goto cleanup;